pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

[tool.pytest.ini_options]
pythonpath = ["."]

[project]
name = "essay-scoring"
version = "0.1.0"
description = "An API to score essays based on a reference answer."
authors = [
    {name = "Gemini"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "mysql-connector-python (>=9.4.0,<10.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "stop-words (>=2018.7.23,<2019.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "xgboost (>=3.0.3,<4.0.0)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
